requirements.txt
Faker==38.2.0
greenlet==3.2.4
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pillow==12.0.0
playwright==1.55.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.2
pytest==9.0.1
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-rerunfailures==16.1
typing_extensions==4.15.0
tzdata==2025.2
